Add move example, add tests, get to working version
The module now has a package.json so it can be included as a module in
other projects successfully.

The ContinuousGenerator.njs (main) file is pretty much complete.. it has
all basic functionality.

I added some "bulletproofing" as they say. Now the continue function can be
called synchronously and everything will be ok. See the comments for why I did
this.